chore(dashboard,ios-agent): lint the whole dashboard source, drop the dead Simulator hide
`packages/dashboard`'s lint script was `eslint src`, and `hooks/`, `components/` and `lib/` are
siblings of `src/` — so 83 of its 123 source files were never linted, while tsconfig type-checked
all four. Widen the script to match tsconfig, and fix everything that surfaced.

hooks (6 findings): render-time ref writes moved into effects (useRelay, useDecoderStream); the
recursive rafLoop became a named function expression; an unused import; one Fast Refresh
suppression with its reason.

components (18 findings): 6 missing `recordCanvasRef` deps — a stable useRef return, so the
effect cadence is unchanged; `load` wrapped in useCallback so its effect dependency is honest
(listing it as-is would have refetched on every render); an unused import; two false positives
suppressed per line with the evidence recorded rather than restructured, since one of them sits
on the Android rotation path whose display behaviour no test covers.

shadcn primitives keep their `*Variants` co-export: `react-refresh/only-export-components` is
waived for `components/ui/**` only. Every correctness rule still applies there — which is what
caught a `Math.random()` during render in sidebar.tsx.

`avatarColors` and the perf-notice policy moved to `lib/`. The Fast Refresh warning was pointing
at a real structure problem: a test was importing a component file to reach one of them.

One finding is left suppressed on purpose. `DeviceViewer` can send a stale `resetMode` after a
reconnect; adding the dep is a behaviour fix, not a lint fix, and is tracked separately in #439.

ios-agent: `SimctlWrapper.boot` ended with an `osascript` call to hide Simulator.app whose
callback swallowed every error. On the only supported Xcode (26.x, per docs/guide/requirements.md)
`simctl boot` does not open Simulator.app, so the call failed with -10006 on every boot and hid
nothing. The stale comment claiming "Xcode 14+ auto-opens Simulator.app" goes with it.

Verified: dashboard lint clean across all 123 files, root lint/typecheck/test green, and the
widened gate was mutation-tested per directory (hooks/components/lib each caught a planted
violation) including that the two per-line suppressions still catch unrelated errors.

Closes #423
Closes #431
