+++ packages/dashboard/components/UserAvatar.tsx
@@ -1,18 +1,4 @@
-const PALETTE_SIZE = 6;
-
-function hashName(name: string): number {
-  let hash = 0;
-  for (const c of name) hash = (hash * 31 + c.charCodeAt(0)) & 0xffffffff;
-  return Math.abs(hash);
-}
-
-export function avatarColors(name: string): { bg: string; fg: string } {
-  const i = (hashName(name) % PALETTE_SIZE) + 1;
-  return {
-    bg: `hsl(var(--avatar-${i}-bg))`,
-    fg: `hsl(var(--avatar-${i}-fg))`,
-  };
-}
+import { avatarColors } from '@/lib/avatar';
 
 type Props = {
   name: string;
