+++ packages/dashboard/hooks/useBreadcrumb.tsx
@@ -1,3 +1,6 @@
+/* eslint-disable react-refresh/only-export-components --
+   The Provider and its consumer hook are one unit; splitting them across two files to satisfy
+   Fast Refresh would spread a 19-line module for no reader benefit. */
 import { createContext, useContext, useMemo, useState, type ReactNode } from 'react'
 
 type ContextValue = {
