+++ packages/dashboard/src/pages/settings/Default.tsx
@@ -23,7 +23,7 @@ import {
 } from '@/components/ui/alert-dialog'
 import { toast } from 'sonner'
 import { Pencil } from 'lucide-react'
-import { avatarColors } from '@/components/UserAvatar'
+import { avatarColors } from '@/lib/avatar'
 import { useAuth } from '@/hooks/useAuth'
 
 type App = { id: number; name: string; bundle_id_key: string; platform: string }
